Add logs and move provider to its own package
